REFACTOR: move the routing logic to the api module
