+++ main.go
@@ -1,33 +1,19 @@
 package main
 
 import (
-	"encoding/json"
 	"fmt"
 	"net/http"
 	"os"
 
-	"github.com/alcb1310/bca-go/models"
-	"github.com/gorilla/handlers"
-	"github.com/gorilla/mux"
+	"github.com/alcb1310/bca-go/api"
 	"github.com/joho/godotenv"
 	"gitlab.com/0x4149/logz"
 )
 
-type response struct {
-	Message string `json:"message"`
-}
-
+// Initialize the logging system with logz
 func init() {
 	logz.VerbosMode()
-
 	logz.Run()
-}
-
-func jsonMiddleware(next http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		w.Header().Add("Content-Type", "application/json")
-		next.ServeHTTP(w, r)
-	})
 }
 
 func main() {
@@ -39,25 +25,9 @@ func main() {
 			logz.Fatal("Unable to load environment variables")
 		}
 	}
-
-	_ = models.Initialize()
-
-	r := mux.NewRouter()
-	r.Use(jsonMiddleware) // All responses will be of type application/json
-
-	r.HandleFunc("/", func(w http.ResponseWriter, r *http.Request) {
-		w.WriteHeader(http.StatusOK)
-		res := response{
-			Message: "Hello World!!!!",
-		}
-		json.NewEncoder(w).Encode(res)
-	})
-
-	originsOk := handlers.AllowedOrigins([]string{"*"})
-	methodsOk := handlers.AllowedMethods([]string{"GET", "POST", "OPTIONS"})
-
-	hand := handlers.CORS(originsOk, methodsOk)(r)
+	r := api.Router{}
+	r.Routes()
 
 	logz.Info("Server Running...\n")
-	logz.Fatal(http.ListenAndServe(fmt.Sprintf(":%s", port), hand))
+	logz.Fatal(http.ListenAndServe(fmt.Sprintf(":%s", port), r.Handler))
 }
